requirements-dev.txt
pytest==2.9.1
docker-py==1.8.0
invoke==0.12.2
recheck==0.2.2
pdbpp==0.8.3
retrying==1.3.3
